requirements.txt
accelerate==0.28.0
gitdb==4.0.11
GitPython==3.1.42
huggingface-hub==0.22.2
numba==0.58.1
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sklearn==0.0.post5
tabulate==0.9.0
tokenizers==0.19.1
torch==2.0.1
torchdistill==0.3.3
torchsummary==1.5.1
torchvision==0.15.2
transformers==4.40.0
wandb==0.16.6
